// models/cutting_sheet.model.ts
export type CuttingSheetState = {
  data: {
    cutt_type: string;
    cutt_finished_size: string;
    cutt_page: string;
    cutt_text_paper: string;
    cutt_printing: string;
    cutt_text_coating: string;
    cutt_500: number;
    cutt_1000: number;
    cutt_2000: number;
    cutt_3000: number;
    cutt_4000: number;
    cutt_5000: number;
  };
  data_text_no: {
    cutt_type: string;
    cutt_finished_size: string;
    cutt_page: string;
    cutt_text_paper: string;
    cutt_printing: string;
    cutt_text_coating: string;
    cutt_500: number;
    cutt_1000: number;
    cutt_2000: number;
    cutt_3000: number;
    cutt_4000: number;
    cutt_5000: number;
  };
  data_text_paper: {
    cutt_type: string;
    cutt_finished_size: string;
    cutt_page: string;
    cutt_text_paper: string;
    cutt_printing: string;
    cutt_text_coating: string;
    cutt_500: number;
    cutt_1000: number;
    cutt_2000: number;
    cutt_3000: number;
    cutt_4000: number;
    cutt_5000: number;
  };
  data_printing: {
    cutt_type: string;
    cutt_finished_size: string;
    cutt_page: string;
    cutt_text_paper: string;
    cutt_printing: string;
    cutt_text_coating: string;
    cutt_500: number;
    cutt_1000: number;
    cutt_2000: number;
    cutt_3000: number;
    cutt_4000: number;
    cutt_5000: number;
  };
  data_cover_paper_edit : {
    cutt_type: string;
    cutt_finished_size: string;
    cutt_page: string;
    cutt_text_paper: string;
    cutt_printing: string;
    cutt_text_coating: string;
    cutt_500: number;
    cutt_1000: number;
    cutt_2000: number;
    cutt_3000: number;
    cutt_4000: number;
    cutt_5000: number;
  }
};